drivers,projects: fix incompatible-pointer-types errors for GCC 14+
GCC 14 promotes -Wincompatible-pointer-types and "makes pointer from
integer without a cast" from warnings to errors, which broke the maxim
builds of many no-OS drivers and example projects. Fix the genuine type
mismatches (not by casting them away):

- Drivers passing a wrong-width pointer to a register accessor now use a
  local of the correct type (e.g. uint16_t/uint8_t/int8_t) and widen
  afterwards, instead of aliasing a uint32_t* as a narrower pointer.
- GPIO reset handling: drivers/examples that assigned a
  no_os_gpio_init_param* into a no_os_gpio_desc* field (or vice versa)
  now acquire the descriptor via no_os_gpio_get[_optional]() or assign
  the correct init-param/descriptor per the API contract.
- iio show/store and debug_reg callbacks corrected to the signatures the
  iio framework expects (intptr_t vs intptr_t*, int vs int32_t return,
  struct iio_device/iio_channel pointer decay).
- no_os_test_bit()/no_os_put_unaligned_be16()/no_os_get_unaligned_be24()
  call sites fixed to pass the buffer/word pointer rather than a value or
  pointer-to-array.
- ade7913/ade9153a driver headers: corrected mistyped struct fields and
  added missing int return types on prototypes.

Builds verified on the maxim platform (ADI arm-none-eabi 15.2 / patched
MSDK) for the affected projects.

+++ projects/lt8722/src/examples/basic/basic_example.c
@@ -47,7 +47,7 @@ int example_main()
 	no_os_uart_stdio(uart_desc);
 
 	int ret;
-	double voltage;
+	int64_t voltage;
 	struct lt8722_dev *lt8722_dev;
 
 	pr_info("\nBasic example");
